Added Key-Value Implementation for validator
This commit includes validating the key values for the data and policy dictionaries. It includes changes to the code, data file, and policy file to ensure that values associated with keys have matching data types. This implementation is not going to be used for the final because it's doing the validation logic rather than passing a data file to the policy file to do the validating.